chore: update transactions query
